chore(internal/stategen): refactor main.go functions to state.go
add state_test.go
